Add package.json to allow installing directly from GitHub
To be clear: this is not supposed to be published to npm or other registries, as the author has previously mentioned to not be willing to publish and maintain a package in the registry. [1](https://github.com/monperrus/crawler-user-agents/issues/14#issuecomment-175269246)

Instead, by simply adding a `package.json` file to the repository, it will allow consumers to install the package directly from GitHub through yarn/npm. For instance:

```
yarn add "https://github.com/monperrus/crawler-user-agents.git#d5ee77dad02fed5d391bb3295355932a9f154071"
# OR
npm install "https://github.com/monperrus/crawler-user-agents.git#d5ee77dad02fed5d391bb3295355932a9f154071"
```

This will not require any future changes from the maintainers, the `package.json` is marked with the `private` flag to prevent publishing to the registry. Users then should explicitly upgrade and review the changes.

I believe this is a good compromise and a viable approach for users, and this does not require further maintainance work.