Add npm dependencies with lockfile
Add pacote (at v11.1.11) as a proper npm dependency because it was being
installed in CI dynamically. It is added as a devDependency because
users of the package do not need it so shouldn't be included for them.
Accordingly, update CI pipelines to use `npm clean-install`, which will
install exactly what is in the lockfile, for more details see:
<https://docs.npmjs.com/cli/v10/commands/npm-ci>.

This change hardens the supply chain. Before, even if you trust npm to
be immutable and always give you the same code for pacote, it would
resolve different versions of transitive dependencies (evidenced by the
presence of version ranges in the dependency list for pacote in the
newly added lockfile, line 1294 to 1314).