Add and use `requirements.txt`
Add a `requirements.txt` file to pin dependencies, direct and
transitive. This was achieved by, in a clean Docker image, first running
`pip3 freeze` to get irrelevant packages, then running the existing
`pip3 install` (from CI) command, and then running `pip3 freeze` again
to get this project's dependencies.

Both CIs have been updated to use the requirements.txt file to install
dependencies. Besides improving reproducibility, this also avoids
duplication.

The benefit of doing this is that the same versions of Python
dependencies will always be used for this project. If the registry is
trusted you can also be sure that the same source code is always run
(the absence of local checksums means this isn't the case if the
registry isn't trusted).